move tests back out of qt
